fix: コードのリファクタ
