+++ mobile/src/features/auth/screens/TutorialStepTwoScreen.tsx
@@ -13,10 +13,11 @@ import {
   TUTORIAL_PROGRESS_FILL_DURATION_MS,
   TUTORIAL_ROUTE_TRANSITION_DELAY_MS,
 } from '@/features/auth/screens/tutorialConfig';
+import { APP_ROUTES } from '@/shared/lib/routes';
 import { useTutorialStore } from '@/shared/stores/tutorialStore';
 
-const NEXT_ROUTE = '/(auth)/tutorial-step-three' as unknown as Href;
-const SKIP_ROUTE = '/(tabs)' as unknown as Href;
+const NEXT_ROUTE = APP_ROUTES.authTutorialStepThree;
+const SKIP_ROUTE = APP_ROUTES.tabs;
 
 export function TutorialStepTwoScreen() {
   const router = useRouter();
