+++ mobile/src/features/session/lib/nativeImagePicker.ts
@@ -1,0 +1,20 @@
+import { NativeModules } from 'react-native';
+
+type ExpoNativeModulesGlobal = typeof globalThis & {
+  expo?: {
+    modules?: Record<string, unknown>;
+  };
+};
+
+type LegacyExpoNativeProxy = {
+  exportedMethods?: Record<string, unknown>;
+};
+
+export function hasNativeImagePickerModule() {
+  const expoModules = (globalThis as ExpoNativeModulesGlobal).expo?.modules;
+  const legacyExpoModules = NativeModules.NativeUnimoduleProxy as LegacyExpoNativeProxy | undefined;
+
+  return Boolean(
+    expoModules?.ExponentImagePicker || legacyExpoModules?.exportedMethods?.ExponentImagePicker,
+  );
+}
